+++ src/main/java/org/albacete/simd/experiments/ExperimentBNBuilder.java
@@ -245,7 +245,7 @@ public class ExperimentBNBuilder {
                 + this.dfmm[1] + ","
                 + this.dfmm[2] + ","
                 + this.nIterations + ","
-                + elapsedTime/1000 + "\n";//this.elapsedTime + "\n";
+                + (double) elapsedTime/1000 + "\n";//this.elapsedTime + "\n";
     }
 
     @Override
